+++ hart/providers/ec2.py
@@ -124,7 +124,7 @@ class EC2Provider(BaseProvider):
             size=None,
             **kwargs):
         if size is None:
-            size = default_size
+            size = self.default_size
 
         zone = kwargs.get('zone')
         if not zone: